refactor(fetcher): transform vaksinid output data to be Contact-compatible
Transforms the output of the vaccination db fetcher to VaccinationContact[].

This will make it ready to be used downstream in the components.

The VaccinationContact interface has a 'map' property.
This is unecessary as WBW's "Buka Peta" button uses the address instead.
Though, I still add it in case we want to use that for the "Buka Peta" button.